refactor(site): extract ProfileDossier into shared component
